requirements.txt
tensorflow>=2.0,<3.0
tensorflow-datasets>=4.0
tensorflow-hub>=0.12
numpy>=1.19
matplotlib>=3.3
pillow>=8.0
h5py>=2.10

